backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
./bedrock-python-sdk/boto3-1.26.142-py3-none-any.whl
./bedrock-python-sdk/botocore-1.29.142-py3-none-any.whl
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
dataclasses-json==0.5.12
Flask==2.3.2
frozenlist==1.4.0
google-api-core==2.11.1
google-auth==2.22.0
google-cloud-aiplatform==1.28.1
google-cloud-bigquery==3.11.4
google-cloud-core==2.3.3
google-cloud-resource-manager==1.10.2
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.2
grpcio-status==1.56.2
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
langchain==0.0.235
langsmith==0.0.7
MarkupSafe==2.1.3
marshmallow==3.19.0
multidict==6.0.4
mypy-extensions==1.0.0
numexpr==2.8.4
numpy==1.24.4
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
proto-plus==1.22.3
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.11
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
s3transfer==0.6.1
Shapely==1.8.5.post1
six==1.16.0
SQLAlchemy==2.0.19
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.7.1
typing-inspect==0.9.0
urllib3==1.26.16
waitress==2.1.2
Werkzeug==2.3.6
yarl==1.9.2
zipp==3.16.2
